# [master < T580-FL] Improve LOAD CSV speed