--- a/src/query/plan/operator.cpp
+++ b/src/query/plan/operator.cpp
@@ -212,7 +212,7 @@ VertexAccessor &CreateLocalVertex(const NodeCreationInfo &node_info, Frame *fram
   // TODO: PropsSetChecked allocates a PropertyValue, make it use context.memory
   // when we update PropertyValue with custom allocator.
 
-  auto start = std::chrono::steady_clock::now();
+  // auto start = std::chrono::steady_clock::now();
   if (const auto *node_info_properties = std::get_if<PropertiesMapList>(&node_info.properties)) {
     for (const auto &[key, value_expression] : *node_info_properties) {
       PropsSetChecked(&new_node, key, value_expression->Accept(evaluator));
@@ -224,14 +224,14 @@ VertexAccessor &CreateLocalVertex(const NodeCreationInfo &node_info, Frame *fram
       PropsSetChecked(&new_node, property_id, value);
     }
   }
-  auto end = std::chrono::steady_clock::now();
-  std::chrono::duration<double> dif = end - start;
-  total += dif;
+  // auto end = std::chrono::steady_clock::now();
+  // std::chrono::duration<double> dif = end - start;
+  // total += dif;
 
-  if (total > new_print) {
-    std::cout << "Time difference props set checked = " << total.count() << "[s]" << std::endl;
-    new_print += static_cast<std::chrono::duration<double>>(10.0);
-  }
+  // if (total > new_print) {
+  //   std::cout << "Time difference props set checked = " << total.count() << "[s]" << std::endl;
+  //   new_print += static_cast<std::chrono::duration<double>>(10.0);
+  // }
 
   (*frame)[node_info.symbol] = new_node;
   return (*frame)[node_info.symbol].ValueVertex();