--- a/src/storage/v2/property_store.cpp
+++ b/src/storage/v2/property_store.cpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Memgraph Ltd.
+// Copyright 2023 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -970,6 +970,8 @@ PropertyValue PropertyStore::GetProperty(PropertyId property) const {
   if (FindSpecificProperty(&reader, property, &value) != DecodeExpectedPropertyStatus::EQUAL) return PropertyValue();
   return value;
 }
+
+PropertyValue PropertyStore::GetEmptyProperty() const { return PropertyValue(); }
 
 bool PropertyStore::HasProperty(PropertyId property) const {
   uint64_t size;