… src/CMakeLists.txt
… src/storage/v2/vertex_accessor.cpp
… src/storage/v2/vertex_accessor.hpp