The idea of this top comment is the following:
- give a quick overview of what was improved (`1. What's improved`)
- give a more detailed overview of benchmarking (`2. Benchmarking`)
- explain steps towards speed improvements found during profiling (`3. Profiling summary`)


## 1. What's improved

After profiling analysis available in section `3. Profiling summary` there were 2 Pull requests to branch `master` of Memgraph:
 - #788 
 - #774

These two PRs proved to **increase** the performance of LOAD CSV import on rich graphs with lots of properties which was shown to be a critical point for improvement.

Another **important** thing to notice is that when doing the import, the **user** can **drastically enhance the performance** of import by using the `READ UNCOMMITED` isolation level on the transaction, due to the fact that it is probably the only transaction doing changes, and such transaction in most cases should be able to see its changes (changes that transaction made).


Description of test: Import X nodes with Y properties each from 1 file (`nodes.csv`) and then `Z` edges with `W` properties from another file( `edges.csv`).

| X | Y | Z| W |Throughput time (OLD)  [node/seconds] | Thorughput time (NEW) nodes [seconds] | Throughput (OLD) [edges/s] | Throughput (NEW) [edges/s] |
| -- | -- | -- |  -- | -- |  -- | -- | -- |
|100000 |10 | 200000| 10 |  120 K/s ( 0.83 s)  | **165 K/s (0.6 s)**  |  70 K/s (2.85 s) | **85 K/s (2.36 s)**  |
|100000 |20 | 200000| 20 |  48 K/s ( 2.06 s)  | **102 K/s (0.96 s)**  |  38 K/s (5.25 s) | **62.5 K/s (3.24 s)**  |
|100000 |30 | 200000| 30 |  28 K/s ( 3.5 s)  | **74 K/s (1.37 s)**  |  15 K/s (7.8 s) | **50 K/s (4.05s)**  |

## 2. Benchmarking

The idea of benchmarking was to check how much we improved with new changes introduced from **Profiling summary** down below, and how different ISOLATION levels on transaction impact throughput

 1. Import nodes and edges from the same file (in the same file there are 2 nodes and 1 edge per line)
- Description: Import 1 000 000 lines, and from each line create 2 nodes and 1 new edge. Each line will have X properties, of 10 string length. Each node has X/2 features, where X is even. No header.

| Number of properties per line | Import time (OLD) [seconds] | Import time (NEW) [seconds] | Throughput (OLD) [lines/s] | Throughput (NEW) [lines/s] |
| -- | -- | -- |  -- | -- |
|2 | 3.6s | 3.917 s| **277K** | 256K |
|10 | 9.6s | 8.4 s|  104 K | **119 K** |
|20 | 22s | 12.9 s| 45.5 K | **77.5 K**  |


2. Import nodes from one file, and then edges from another file
- Description: Import X nodes with Y properties nodes from 1 file and then Z edges with W properties from another file.

| X | Y | Z| W |Throughput time (OLD)  [node/seconds] | Thorughput time (NEW) nodes [seconds] | Throughput (OLD) [edges/s] | Throughput (NEW) [edges/s] |
| -- | -- | -- |  -- | -- |  -- | -- | -- |
|1M |10 | 5M| 5 |  117K/s (total 8.8seconds)  |  | 51K/s (total: 98 seconds)  |   |

3. Import nodes from one file, and edges from another file using `SNAPSHOT` or `READ UNCOMMITTED` isolation level
- Description: Import X nodes with few properties [1,3] and Z edges with few properties [1,3]

| X | Z |Throughput time (`SNAPSHOT`)  [node/seconds] | Throughput time (`READ UNCOMMITTED`) nodes [seconds] | Throughput (`SNAPSHOT`) [edges/s] | Throughput (`READ UNCOMMITTED`) [edges/s] | ratio (edges/nodes)|
| -- | -- | -- |  -- | -- | -- | -- |
|3500 | 50K|  50K/s (0.07 seconds)  |  50K/s (0.07 seconds)  | 100K/s (0.5 seconds) | 100K/s (0.5 seconds) | 14 |
| 3500 | 500K|  50K/s (0.07 seconds)  |  50K/s (0.07 seconds)  | 16K/s (31 seconds) | 106 K/s (4.7 seconds) :heavy_check_mark:  |  **142** |
|3.5M | 5M|  397 K/s (8.8 seconds)  |  397 K/s (8.8 seconds) | 90 K/s (55.1 seconds) |  92 K/s (54 seconds) | 1.5|




## 3. Profiling Summary

**The Problem and Goal**
LOAD CSV speed: 5000 lines/second at graphs with a lot of properties -> get at least 10 000 lines/second :heavy_check_mark: 

**Context**
 Dataset + Queries = Workload -> Read **Profiling Summary** below

| Experiment Name | Description | Throughput (lines/s) |  Throughput (edges/s) |
| -- | -- | -- | -- |
| Check baseline import speed on 100 000 lines, with 100 string properties per line, 10 char each string | Read more description [here](https://github.com/memgraph/memgraph/pull/753#issuecomment-1411921127).  **The test was done on [1]** | 4350 | 4350 |
| Implement `ReferenceExpressionEvaluator` to avoid copying every `row` when evaluating the expression | Visitor was the real bottleneck and I confirmed this with `gperftools` and by [measuring time spent in that function](https://github.com/memgraph/memgraph/pull/753#issuecomment-1411921127), it made sense to avoid copying the whole list just to get one object from it. **The test was done on [1]** | 8550 | 8550 |
|  Lookup of parameters [takes time](https://github.com/memgraph/memgraph/pull/753#issuecomment-1413677072) as it needs to iterate through `std::vector` each time to find a specific index | After profiling, it seemed it is better with `ReferenceExpressionEvaluator`, but `AtTokenPostion` was still taking a lot of time. I changed std::vector with std::map for faster lookup, in this case. **The test was done on [1]** | 9520 | 9520 |
| Test batch insertion of properties when creating vertices | Test batch insert of properties due to the fact that [flamegraph](https://user-images.githubusercontent.com/61245998/216326788-a423acc3-f4b4-473f-ba44-4378e0756529.svg) showed (read more [here](https://github.com/memgraph/memgraph/pull/753#issuecomment-1415948417)) it took a lot of time to insert property 1 by one. Due to guarantees on creating vertex that no two of the same properties can exist inside `{}`, we can batch insert properties without checking if it exists or not. **The test was done on [1]** | 10.9K | 10.9K |
| Implement `SetProperties` in `PropertyStore`| From profiling in flamegraph, it seems a lot of time is wasted on the DecodeExpectedProperty function which shows that probably problem is that for each new property when there is the setting of a new property, we need to check whether this property exists already, and copy memory block of properties. To avoid that, we can set all properties all at once. **The test was done on [1]** | 18.1K | 18.1K |
| Test using std::vector instead of `std::map` to process properties in batch| From flamegraph it seems like we can squeeze out a little bit more time out of execution since std::map takes a bit of time. **The test was done on [1]** |**>20K/s** | **>20K/s** |

**Test Summary**
| Test ID | Test query | Test description| Test file|
|--| --| --|--|
| 1 | `LOAD CSV FROM ".../generated_text.csv" NO HEADER AS row CREATE (a: Label {arg0:row[0],...,arg49:row[49]}), (b: Label {arg50:row[50],...,arg99:row[99]}), (a)-[:CONNECTED_TO]->(b);`  | 100 000 lines file with 100 properties per line, 10 char string each property. Create 2 nodes and 1 edge between them for each line (200 000 nodes created and 100 000 edges)| |
|TODO | TODO | TODO | TODO |