@antoniofilipovic 2023-01-30
- [ ] Test with 100 000 lines with 50 props per node and without Lab, create 2 nodes and 1 edge per line
```
LOAD CSV FROM "/home/antonio/work/memgraph/github-repos/core/generated_text.csv" NO HEADER AS row
-> CREATE (a: Label {arg0:row[0],...,arg49:row[49]}),
-> (b: Label {arg50:row[50],...,arg99:row[99]}),
-> (a)-[:CONNECTED_TO]->(b);
```
Result:
![dump-csv-100000-lines-node-props-50-string-15-edge-props-0-no-lab](https://user-images.githubusercontent.com/61245998/215497700-eea3ddce-7f31-41ed-ab8f-9e3b608c9160.svg)