@antoniofilipovic 2023-01-30
- [ ] Test reading 100 000 lines with 50 props per node, create 2 nodes and 1 edge only for the last line:-> Empty set (14.435 sec)

![dump-csv-100000-lines-node-props-50-string-15-edge-props-0-no-lab-last-line-only](https://user-images.githubusercontent.com/61245998/215500570-40c821c2-0dd6-4608-8379-be40f631cdf5.svg)