@antoniofilipovic 2023-01-30
Seems like problem is node creation, since reading through file is extremely fast, will test reading in Release mode, as all these two tests were done in Deubg mode