@gitbuda 2023-01-30
> reading the whole file and creating 2 nodes and 1 edge only for the last line is really fast. Release mode: **(1.383 sec)**

Yes, reading is fast, the problem is Memgraph single thread speed -> 2 threads one for reading and one for execution (one more thread for reading) won't be very beneficial because parsing is ~2% of the execution?