@antoniofilipovic 2023-01-30
- [ ] Test with 100 000 lines, 50 props per node without Lab, only create nodes and edges, no properties

16.9 seconds to create 200 000 nodes and 100 000 edges and read 100 000 lines but create 0 properties in Debug mode

Link [here](https://user-images.githubusercontent.com/61245998/215533727-89fccd3d-266d-41bd-8b73-1a9583c15fc6.svg)