@antoniofilipovic 2023-01-31
Measuring time of function:
`Result<PropertyValue> VertexAccessor::SetProperty(PropertyId property, const PropertyValue &value)`

Out of 299 seconds to execute, 120 seconds are spent in above function:


```
Result<PropertyValue> VertexAccessor::SetProperty(PropertyId property, const PropertyValue &value) {
  auto start = std::chrono::steady_clock::now();

  utils::MemoryTracker::OutOfMemoryExceptionEnabler oom_exception;
  std::lock_guard<utils::SpinLock> guard(vertex_->lock);

  if (!PrepareForWrite(transaction_, vertex_)) return Error::SERIALIZATION_ERROR;

  if (vertex_->deleted) return Error::DELETED_OBJECT;

  auto current_value = vertex_->properties.GetProperty(property);
  CreateAndLinkDelta(transaction_, vertex_, Delta::SetPropertyTag(), property, current_value);
  vertex_->properties.SetProperty(property, value);

  UpdateOnSetProperty(indices_, property, value, vertex_, *transaction_);

  auto end = std::chrono::steady_clock::now();
  std::chrono::duration<double> dif = end-start;
  total+=dif;

  if (total>new_print){
    std::cout << "Time difference = " <<  total.count() << "[s]" << std::endl;
    new_print+=static_cast<std::chrono::duration<double>>(10.0);
  }

  return std::move(current_value);
}
```