@antoniofilipovic 2023-01-31
Next step is to measure where is around 140 seconds spent if it is not in `SetProperty`