@antoniofilipovic 2023-01-31
The next hypothesis is that there is additional time spent on `visitor`. To test that, I set a timer in function:
`TypedValue Visit(SubscriptOperator &list_indexing) override`

I run the following test:
- [ ]  Test on CSV file with 100 000 lines with 50 props per node and without Lab, create 2 nodes and 1 edge per line: (270 sec)

In this test, the time spent in the function `Visit(SubscriptOperator &list_indexing)` is 100 seconds.

Summary:

- out of 270 seconds in total, 100 is spent in `Visit(SubscriptOperator &list_indexing)` 
- out of 270 seconds in total, 100 is spent in `SetProperty` function
- out of 270 seconds, 240 seconds is by calling `PropsSetChecked` function