@antoniofilipovic 2023-01-31
The next step is to confirm if this Visit is a real bottleneck and debug in `RelWithDebInfo` mode.