@antoniofilipovic 2023-02-01
The latest hypothesis is that there is additional time spent on `visitor`. To test that, I set a timer in function:
`TypedValue Visit(SubscriptOperator &list_indexing) override ` and built Memgraph in `RelWithDebInfo`.

Testing hypothesis:
- [ ] Test on CSV file with 100 000 lines with 50 props per node and without Lab, create 2 nodes and 1 edge per line: (23.3 sec)

Out of 23.3 seconds in execution, 7 seconds were spent in `visitor`. That accounts for 30% of the time.