@antoniofilipovic 2023-02-01
The next step is to implement reference-based expression eval to avoid copying of the list.