@antoniofilipovic 2023-02-01
Another idea is to serialize also all properties at once, so we don't lose much time there. This is not the immediate next step, just one of the ideas to test afterward.