@antoniofilipovic 2023-02-01
In the latest commit, I implemented `ReferenceExpressionEvaluator` to get from frame pointer to list object, without copying it.

I tested now hypothesis that this will speed up going through CSV:

- [ ] Test on CSV file with 100 000 lines with 50 props per node and without Lab, create 2 nodes and 1 edge per line: (11.7 sec)

Total time was reduced by **half**. It now "reads" 10k/s. Instead of 5k/s.