@antoniofilipovic 2023-02-02
Next step is to profile again if it seems better. I profiled this time with `flamegraph` and here is the result:

![flamegraph](https://user-images.githubusercontent.com/61245998/216313339-aaaa3c14-698e-4146-b21d-210226b0e939.svg)

It seems like `Vistor` is still very slow and the property is taking an equal amount of time. I will try to profile again with gperftools current state, and also profile with `flamegraph` before `ReferenceExpressionEvaluator` was introduced.