@antoniofilipovic 2023-02-02
Hypothesis: lookup of parameters takes some time as it needs to iterate through std::vector each time to find a specific id.

- [ ] Test and profile introducing std::unordered_map instead of vector on  CSV file with 100 000 lines with 50 props per node and without Lab, create 2 nodes and 1 edge per line

![flamegraph](https://user-images.githubusercontent.com/61245998/216326788-a423acc3-f4b4-473f-ba44-4378e0756529.svg)


Result: reduced time of import by **at least 1 second**. Now time is 10.5 seconds in most cases instead of 11.7 seconds.