@antoniofilipovic 2023-02-02
The next step is to reduce time by setting properties as a batch instead of one by one.