@antoniofilipovic 2023-02-02
One thing I tested.

What if two same properties appear inside `{}` like `arg0` in case below:

```
LOAD CSV FROM ".../generated_text.csv" NO HEADER AS row
CREATE (a: Label {arg0:row[0],arg1:row[1],arg0:row[2]}),
(b: Label {arg50:row[50],arg51:row[51],arg50:row[52]}),
(a)-[:CONNECTED_TO]->(b);
```

It throws an error:
`Same key can't appear twice in a map literal.`

This is why it  should be possible to optimize to batch process as we don't need to check for each property if it appeared already, only set new property.

On the other hand following situation is not possible currently to optimize:
```
LOAD CSV FROM "...generated_text.csv" NO HEADER AS row
CREATE (a: Label),
(b: Label),
(a)-[:CONNECTED_TO]->(b)
SET a.arg0=row[0], 
		a.arg1=row[1],
		a.arg0=row[2];
```
Because it generates multiple `SetProperty` cursors, and we can't batch process in that case.