@antoniofilipovic 2023-02-20
1. Import X nodes with Y properties nodes from 1 file and then Z edges with W properties from another file.

| X | Y | Z| W |Throughput time (OLD)  [node/seconds] | Thorughput time (NEW) nodes [seconds] | Throughput (OLD) [edges/s] | Throughput (NEW) [edges/s] |
| -- | -- | -- |  -- | -- |  -- | -- | -- |
|1M |10 | 5M| 5 |  117K/s (total 8.7seconds)  | TODO | 61.5 K/s (total: 81.3 seconds)  | TODO  |

Benchmarked again today, and it seems performance was much better as there was no need to activate the swap.